Regroups namespaces related tests
